Fixes #2433
